django_report/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.3
djangorestframework==3.14.0
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
numpy==1.23.5
packaging==21.3
pandas==1.5.2
prompt-toolkit==3.0.33
psycopg2==2.9.5
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
redis==4.3.5
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.5
